All collections made either List or Set
It’s good coding practice to code against the interface an not the
implementation.

It’s almost always preferable to code against the interface and not the
implementation. This way, if you need to change the implementation
later, it won’t break consumers who code against the interface - Matt
Ball

Returning List will make it possible for users of your library to use
anything that implements a List interface, while using an ArrayList
will force them to use an ArrayList.- ipvalic